+++ wormhole-connect-loader/package.json
@@ -3,7 +3,11 @@
   "version": "0.0.1-beta.0",
   "private": true,
   "dependencies": {
-    "@mui/material": "^5.12.1",
+    "@mui/material": "^5.12.1"
+  },
+  "devDependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -11,23 +15,18 @@
     "@types/node": "^16.18.23",
     "@types/react": "^18.0.37",
     "@types/react-dom": "^18.0.11",
-    "react-scripts": "5.0.1",
+    "babel-loader": "^9.1.2",
+    "ts-loader": "^9.4.2",
     "typescript": "^4.9.5",
-    "web-vitals": "^2.1.4"
-  },
-  "devDependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "webpack": "^5.79.0",
+    "webpack-cli": "^5.0.1"
   },
   "peerDependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "build": "tsc"
   },
   "eslintConfig": {
     "extends": [
@@ -49,6 +48,8 @@
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
   "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [ "lib" ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/wormhole-foundation/wormhole-connect.git"
